backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
djangorestframework-simplejwt>=5.3,<6.0

celery>=5.4,<6.0
requests>=2.32,<3.0
stripe>=10.0,<11.0

django-environ>=0.11,<1.0
PyMySQL>=1.1,<2.0

gunicorn>=22.0,<23.0
